[#138] Enforce order on the IATI Schemas
This replaces all choice elements in the activity, organisation and
common schemas with sequence elements.

Note that in this commit almost no constraints are placed on how many
times elements can occur. (The only constraint is that iati-identifier
is required, in order to make the "empty should fail" tests pass).

More constraints will be added in a later commit.