[#164] Fixes many tests that were failing for the wrong reason
This fix deals with many of the glaring problems, mainly left over
from version 1.x tests.
It also fixes some of the remaining problems that occurred with a
late re-arrangement of the schema elements
I think anything left over now is going to be accidental, copy and
paste problems.
I believe most things in activity-tests/should fail are doing so for
the right reasons. (maybe!)