Remove test files that are valid against the Schema
The 2.02 Schema has no checks to ensure attribute values are not empty.
As such, test files that require this assumption are valid when they
should not be.

The test files have been removed since they pass despite being 'should-fail'
tests.

It may be wondered why they got into 2.02 despite actually passing. This
is because there was another error with the files that meant they were not
actually XML. This was fixed in:
  a31d8b6adc33cdf2a539288ad9d7e2e5b94c2f8d