Update country and region rule - 2.03
The rules within the IATI Standard and the surrounding documentation are unclear and inconsistent whether or not publishers can map their activities to different region vocabularies. Or, if all reported regions + countries have to make up 100% of the activity.

We propose a bug fix to confirm that publishers can map their activities to multiple region vocabularies e.g. 100% to OECD DAC and 100% to UN Region list.

https://github.com/IATI/IATI-Extra-Documentation/pull/516